+++ frontend/srcs/src/app/components/friend-list/friend-sidebar/friend-sidebar.component.ts
@@ -1,10 +1,11 @@
 import { Component, Input } from '@angular/core';
-import { ApiService, AuthService } from 'src/app/services';
+import { ApiService, AuthService, GameService } from 'src/app/services';
 import { UsersCache } from 'src/app/cache';
 import { BaseComponent, FriendListComponent } from 'src/app/modules';
 import { User } from 'src/app/models';
 import { UriConstants } from 'src/app/utils';
 import { NavigationExtras, Router } from '@angular/router';
+import { ChallengeService } from 'src/app/services/challenge.service';
 
 @Component({
   selector: 'app-friend-sidebar',
@@ -23,6 +24,9 @@ export class FriendSidebarComponent extends BaseComponent<User> {
     private readonly authService: AuthService,
     private readonly parent: FriendListComponent,
     private readonly router: Router,
+    private readonly gameService: GameService,
+    private readonly challengeService: ChallengeService
+
 
     ) {
       super(api);
@@ -72,4 +76,38 @@ export class FriendSidebarComponent extends BaseComponent<User> {
     this.router.navigate(['/profile/', user.username], navigationExtras);
   }
 
+
+  challengeUserid(userId: string){
+    this.apiService.getService({
+        url: `${UriConstants.CHALLENGE}/${userId}`,
+      }).subscribe({
+        next: () => {
+          this.challengeService.sendChallengingUserIdSub(userId);
+          },
+          error: error => {
+            this.alertConfiguration('ERROR', error);
+            this.openAlert();
+            this.loading = true;
+          },
+    });
+  }
+
+  isOnline(userId: string): boolean {
+    return this.cachedUsers.isUserConnected(userId);
+  }
+
+  canSpectate(userId: string): boolean {
+    return this.cachedUsers.isUserPlaying(userId);
+  }
+
+  spectateGame(userId: string){
+    const gameId = this.cachedUsers.getLiveGameId(userId);
+
+    const navigationExtras: NavigationExtras = {
+      state: { data: { spectate: true  } }
+    };
+
+    this.gameService.sendWannaWatch(Number(gameId));
+    this.router.navigate(["/game/spectate"], navigationExtras);
+  }
 }